requirements.txt
certifi==2021.5.30
click==8.0.4
cmake==3.25.2
colorama==0.4.5
dataclasses==0.8
dlib @ https://pypi.python.org/packages/da/06/bd3e241c4eb0a662914b3b4875fc52dd176a9db0d4a2c915ac2ad8800e9e/dlib-19.7.0-cp36-cp36m-win_amd64.whl#md5=b7330a5b2d46420343fbed5df69e6a3f
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.3
importlib-metadata==4.8.3
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.19.5
opencv-python==3.4.13.47
Pillow==8.4.0
typing_extensions==4.1.1
Werkzeug==2.0.3
wincertstore==0.2
zipp==3.6.0
